# Print JSON format and allow disabling tests skipping